fix(#844) Keep header layout when errors occurred in dataset view
see #844
This PR includes the restructuring of dataset view and new app-header component for better control of header behaviour
