+++ frontend/components/commons/TaskSearch.vue
@@ -16,34 +16,16 @@
   -->
 
 <template>
-  <div
-    :class="[
-      'app',
-      currentTask,
-      annotationEnabled ? '--annotation' : '--exploration',
-      areMetricsVisible ? '--metrics' : null,
-    ]"
-  >
-    <div class="app__content">
-      <section id="header" ref="header" class="header">
-        <ReTopbarBrand v-if="currentTask">
-          <ReBreadcrumbs :breadcrumbs="breadcrumbs" :copy-button="true" />
-          <user />
-        </ReTopbarBrand>
-        <task-sidebar :dataset="dataset" />
-        <component :is="currentTaskHeader" :dataset="dataset" />
-      </section>
-      <div :class="['grid', annotationEnabled ? 'grid--editable' : '']">
-        <Results :dataset="dataset" />
-      </div>
+  <div class="app__content">
+    <div :class="['grid', annotationEnabled ? 'grid--editable' : '']">
+      <Results :dataset="dataset" />
     </div>
   </div>
 </template>
 <script>
 import "assets/icons/copy";
 import { mapActions } from "vuex";
-import { DatasetViewSettings } from "@/models/DatasetViewSettings";
-import { currentWorkspace, workspaceHome } from "@/models/Workspace";
+import { currentWorkspace } from "@/models/Workspace";
 export default {
   props: {
     dataset: {
@@ -51,30 +33,9 @@ export default {
       required: true,
     },
   },
-  data: () => ({
-    headerHeight: undefined,
-  }),
   computed: {
-    currentTask() {
-      return this.dataset.task;
-    },
-    breadcrumbs() {
-      return [
-        { link: { path: workspaceHome(this.workspace) }, name: this.workspace },
-        { link: this.$route.fullPath, name: this.dataset.name },
-      ];
-    },
-    areMetricsVisible() {
-      return this.dataset.viewSettings.visibleMetrics;
-    },
-    currentTaskHeader() {
-      return this.currentTask + "Header";
-    },
     annotationEnabled() {
       return this.dataset.viewSettings.viewMode === "annotate";
-    },
-    globalHeaderHeight() {
-      return this.dataset.viewSettings.headerHeight;
     },
     currentViewMode() {
       return this.dataset.viewSettings.viewMode;
@@ -84,41 +45,17 @@ export default {
     },
   },
   watch: {
-    globalHeaderHeight() {
-      if (this.globalHeaderHeight !== this.headerHeight) {
-        this.headerHeightUpdate();
-      }
-    },
     async currentViewMode(n) {
       if (n === "labelling-rules") {
         await this.resetSearch({ dataset: this.dataset });
       }
     },
   },
-  mounted() {
-    this.setHeaderHeight();
-  },
   methods: {
     ...mapActions({
       resetSearch: "entities/datasets/resetSearch",
       fetchDataset: "entities/datasets/fetchByName",
     }),
-    async setHeaderHeight() {
-      const header = this.$refs.header;
-      const resize_ob = new ResizeObserver(() => {
-        this.headerHeight = header.offsetHeight;
-        this.headerHeightUpdate();
-      });
-      resize_ob.observe(header);
-    },
-    headerHeightUpdate() {
-      DatasetViewSettings.update({
-        where: this.dataset.name,
-        data: {
-          headerHeight: this.headerHeight,
-        },
-      });
-    },
   },
 };
 </script>
@@ -157,73 +94,6 @@ export default {
   z-index: 0;
   .--fixed:not(.fixed-header) & {
     z-index: 2;
-  }
-}
-
-.header {
-  opacity: 1;
-  position: relative;
-  transition: none;
-  top: 0;
-  right: 0;
-  left: 0;
-  transform: translateY(0);
-  position: fixed;
-  background: $bg;
-  z-index: 1;
-  ::v-deep .header__filters {
-    position: relative;
-  }
-  .fixed-header & {
-    animation: header-fixed 0.3s ease-in-out;
-    z-index: 2;
-    box-shadow: 1px 1px 6px $font-medium-color;
-    ::v-deep .filters__title,
-    ::v-deep .topbar {
-      display: none;
-    }
-    ::v-deep .global-actions {
-      margin-top: 0;
-      padding-top: 0;
-      background: $bg;
-      border: none;
-      min-height: 60px;
-    }
-  }
-  .fixed-header .--annotation & {
-    ::v-deep .filters__area {
-      display: none;
-    }
-  }
-}
-
-@keyframes header-fixed {
-  0% {
-    transform: translateY(-150px);
-  }
-  100% {
-    transform: translateY(0);
-  }
-}
-
-.switch-button {
-  border: none;
-  background: lighten($primary-color, 20%);
-  color: $lighter-color;
-  outline: none;
-  padding: 0.5em 1em;
-  cursor: pointer;
-  &:last-of-type {
-    margin-right: 2em;
-    border-top-right-radius: 5px;
-    border-bottom-right-radius: 5px;
-  }
-  &:first-of-type {
-    border-top-left-radius: 5px;
-    border-bottom-left-radius: 5px;
-  }
-  &.selected {
-    background: $primary-color;
   }
 }
 </style>
